requirements.txt
asgiref==3.2.10
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
